UPDATE .env with latest env-example
New configuration values were added to env-example and our old .env was
throwing errors when I started it up. So merged in the new config
values.

Note that a new value, WORKSPACE_VUE_CLI_SERVE_HOST_PORT had a default
value of 8080, this conflicted with Adminer which also wanted to be
served on 8080. I changed the WORKSPACE_VUE_CLI_SERVE_HOST_PORT to 8081
instead.