ADD lti 1.3 example tool
We need a tool to test the shim against. Since the RI tool wasn't
working, we're using the LTI 1.3 PHP Example Tool.

A few changes were made from the original example tool so that it'll
work with the shim's seeded data out of the box.

composer.json - github api is rate limited, so we want composer to pull
from github using https instead, otherwise, rebuilds will eventually
fail if we do too many of them.

Dockerfile - was modified in order to try to track down the issue above,
keeping the changes because it should allow the container to start up
faster. We don't need to mount the src/ directory since we're not
working on the example tool.

src/db/configs/local.json - default config values that should work with
the seeded data in the shim.