ADD ltijs-demo-server, LTI tool to use for testing
The LTI 1.3 PHP Example Tool has proven... problematic... in many areas.
So I'm very happy to be able to replace it with the new Ltijs demo app.
It supports all the services we want to use without us having to modify
the source and also has a nice simple UI and logs that are far easier to
debug. Also really helps that it tells you when requests encounter a
server error, instead of just an unhelpful blank page.

The ltijs-demo-server runs on port 4000, which none of the existing
services seem to run on. It requires a mongo db server, which I've also
added.

I've modified index.js to add the shim as a platform entry. I've also
added Dockerfile and corresponding docker-compose.yml for debugging use.
This is based on the ltijs-demo-server repo with the latest commit on
2020-07-29.

OIDC Login: http://localhost:4000/login
Auth Resp: http://localhost:4000/
Target Link URI: http://localhost:4000/
JWKS (within laradock): http://ltijs-demo-server:4000/keys
ISS: http://localhost:4000/