ADD createdb.sh in postgres to create test db
Laradock's sqlite is too old (3.22.0), we need at least 3.24.0 to
support the newly added upsert() database method in Laravel. So we're
switching to running the tests on the postgres database. The createdb.sh
script will create the test database for us.

Note that createdb.sh doesn't seem to be run if you have existing
postgres data. I had to reset all postgres data (deleting
~/.laradock/data/postgres/) before bringing up the postgres image to get
the test database to be created.